grsec: Apply grsecurity-2.2.0-2.6.34.6-201009041352.patch
Signature-tree: e7d8cfd974de9b49c1863e1652559ded03a22a2e

+++ kernel/relay.c
@@ -1293,7 +1293,7 @@ static ssize_t subbuf_splice_actor(struct file *in,
 		return 0;
 
 	ret = *nonpad_ret = splice_to_pipe(pipe, &spd);
-	if ((int)ret < 0 || ret < total_len)
+	if (ret < 0 || ret < total_len)
 		return ret;
 
         if (read_start + ret == nonpad_end)
